LayoutTests/accessibility/mac/aria-menubar-expected.txt

menubar
This tests that an ARIA menu bar is exposed as a menu bar.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS accessibilityController.focusedElement.role is 'AXRole: AXMenuBar'
PASS successfullyParsed is true

TEST COMPLETE
